% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum-geom-series.R
\name{sum_geom_series}
\alias{sum_geom_series}
\title{Sum a infinite geomertic series}
\usage{
sum_geom_series(a, r)
}
\arguments{
\item{a}{The first value of the series}

\item{r}{The ratio of the subsequent terms of the series.}
}
\description{
This function calculates the sum of a geometric series.
}
